GREENCHAIN-main/greenchain/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
dedalus-labs
anthropic>=0.39.0
python-dotenv==1.0.1
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.1.1
joblib==1.4.2
httpx==0.27.2
beautifulsoup4==4.12.3
eval_type_backport==0.2.2
